Update ProcessBuilder command execution in ProxyConfigurator
Revised the ProcessBuilder command execution in the startProxy() method of the ProxyConfigurator.java class. Previously, it was wrapped within the "/bin/bash", "-c" parameters which was found unnecessary and could potentially result in execution errors. The aim of this change is to increase command execution reliability and simplification of the code.