Update Tor proxy startup command in ProxyConfigurator
Changed the command line used to start the Tor proxy in the startProxy() method in ProxyConfigurator.java. Prior command was using a relative path, which could lead to issues when running the command from different directories. The start command now uses an absolute path to the 'torrc-proxy' file to prevent such issues.