"Add functionality for file uploading, refactor Nginx config creation"
This commit was made to introduce a functionality that allows multiple files to be uploaded at once. This is done through a new POST mapping in the OnionServiceController class. As a result, a new HTML form for file uploading was also created.

Also, there was a refactor in the buildNginxConfig() method, with Nginx configuration being formatted using String.format() to enhance readability. The method writeNginxConfig() was removed, as its functionality got replaced by the uploadFiles() method.

Unused imports have also been removed, and new necessary ones (such as MultipartFile) have been added.