"Introduce file upload functionality to OnionServiceController"
This commit introduces the ability to upload multiple files simultaneously through a new POST mapping in the OnionServiceController class. This addition necessitated the creation of an HTML form for file uploads.

This commit also includes a refactor of the buildNginxConfig() method. The Nginx configuration is now formatted using String.format() to improve readability and facilitate maintenance. As a result, the writeNginxConfig() has been removed, as its functionality is now handled by the uploadFiles() method.

Lastly, unused import statements were removed and necessary ones (such as MultipartFile) were added.