"Update file upload path to fixed location"
Changed the file upload directory in OnionServiceController from dynamic to a fixed path. This simplifies the configuration by eliminating the need to dynamically change upload paths based on hidden service port values. This makes sense since all services are currently using the default port 80. This may need to be revised if the system is updated to support multiple ports.