Remove test page, add file deletion feature in UI
Deleted the Test Onion Service page from the codebase as its purpose was served. In addition, a new feature has been added that displays all of the uploaded files on the file upload page and provides an option to delete any of these files. This has primarily been done to enhance the user interface and offer more file management options to the users. Changes affected the file_upload_form.html template and the corresponding OnionServiceController.java with additions of dedicated methods for file removal and list retrieval.