Improve relay configuration by removing SOCKSPort
The 'SOCKSPort' setting was removed in RelayConfig, BaseRelayConfig models, and also in TorConfigurationService. This was done since handling of the 'SocksPort' is unnecessary for the current operational logic of the application. The line-checking for 'SocksPort' was also removed in the file reader in RelayUtils, and the related form field was excluded from relay-config.html.