Remove unused SocksPort setting in BridgeController
The "SocksPort 0" configuration was removed from BridgeController.java as it is not relevant in the context of a Tor bridge configuration. The setting was causing confusion because it's specifically meant for client configurations but was placed in the BridgeController. Consequently, keeping the code clean and easier to understand, this commit removes this unnecessary configuration parameter.