Add support for IPv6 in TorrcFileCreator and corresponding tests
Expanded the TorrcFileCreator class to support IPv6, writing the result to the Tor configuration file if the system supports IPv6 and is reachable. This feature was implemented by adding a new method called getSystemIpv6() which retrieves the system's IPv6.
The TorrcFileCreatorTest was also created for proper testing of the changes, including mocking the InetAddress class. In addition, a new mockito-inline dependency was added to the pom file for testing purposes.