Update API endpoint URLs to localhost
This commit updates the base API URL and subsequent endpoint URLs in the data.js file
from the static '192.168.2.118' to localhost '127.0.0.1'.
This is to allow the application to run locally without the need for a hard-coded,
fixed IP address. Hence, improving the flexibility and portability of the application
for different development environments.