Refactor install.sh for custom nginx configuration
Updated the install.sh script to install components individually, and if the component is nginx, the script will also modify the nginx.conf file. Instead of installing the components all at once, they are now installed one by one to allow for custom configurations when certain components (such as nginx) are installed.