Update user determination method in install.sh
Changed the method of determining the executing user from 'whoami' to 'logname' in the install.sh script. This change will allow the script to identify the user who initiated the script rather than the user currently executing it, which is particularly important when modifying the nginx.conf file in component-specific configurations.